+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to Toisto will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v0,0.12 - unreleased
+
+### Added
+
+- Added a nature topic.
+- Added shopping center to the city topic.
+
 ## v0.0.11 - 2022-10-27
 
 ### Added